FIX: reconcile recreating records that already exist (#15)
https://dev.discourse.org/t/-/39001

* FIX: reconcile publishes PTR and TXT records unconditionally

Looking at metrics it appears that new records are being published
every time reconcile is done (currently every 5 minutes).

DDNSSD::Backend publish_record will record a op: "publish",
which ends up in the ResourceRecordStore which chooses not to do an
INSERT because it finds that an identical record already exists.
But it gets recorded in prometheus anyway.

So if we can see that the PTR and TXT records already exist, don't
publish them again.

* FIX: removing host record and then adding it again during reconcile

Reconcile was always removing the host record. Immediately after
reconcile is done, the host record was being inserted again.